Code/requirement.txt
absl-py==1.0.0
attrs==21.4.0
auraloss==0.2.1
backcall==0.2.0
bokeh==2.4.2
cachetools==4.2.4
cdpam==0.0.6
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
configparser==5.2.0
debugpy==1.4.1
decorator==5.0.9
Deprecated==1.2.13
dill==0.3.4
docker-pycreds==0.4.0
featureflow==3.0.3
filelock==3.4.2
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
importlib-metadata==4.10.0
importlib-resources==5.4.0
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==6.1.12
jupyter-core==4.7.1
linecache2==1.0.0
llvmlite==0.37.0
lmdb==1.3.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
msgpack==1.0.3
munkres==1.1.4
natsort==7.0.1
noisereduce==2.0.1
nose==1.3.7
oauthlib==3.1.1
olefile==0.46
onnx==1.10.2
pandas==1.3.5
parso==0.8.2
pathtools==0.1.2
pickleshare==0.7.5
Pillow==8.4.0
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.19.3
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
Pygments==2.9.0
pyrsistent==0.18.1
PySoundFile==0.9.0.post1
python-dateutil==2.8.2
pytz==2022.1
pywin32==301
PyYAML==6.0
pyzmq==22.1.0
ray==1.9.2
redis==4.1.2
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.8
sentry-sdk==1.5.2
shortuuid==1.0.8
smmap==5.0.0
snorkel==0.9.8
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
termcolor==1.1.0
torch==1.10.1
torchaudio==0.10.1
torchinfo==1.6.3
torchvision==0.11.2
tornado==4.5.3
tqdm==4.64.0
traceback2==1.4.0
traitlets==5.0.5
ujson==5.2.0
unittest2==1.1.0
wandb==0.12.9
wcwidth==0.2.5
Werkzeug==2.0.2
wget==3.2
wincertstore==0.2
wrapt==1.13.3
yaspin==2.1.0
zipp==3.5.0
